Confirm on delete

+++ ui-react-native-web/src/components/RecordList.js
@@ -1,7 +1,7 @@
 import React, { Fragment, useState } from 'react';
 import { Text, View, StyleSheet, Switch, FlatList, Button } from 'react-native';
 import moment from 'moment';
-import Constants from "../constants";
+import Constants from '../constants';
 
 const styles = StyleSheet.create({
   item: {
@@ -13,7 +13,7 @@ const styles = StyleSheet.create({
   },
 });
 
-const RecordList = ({ records }) => {
+const RecordList = ({ records, onUpdate }) => {
   const [showAll, setShowAll] = useState(false);
 
   const displayRecords = showAll
@@ -27,40 +27,43 @@ const RecordList = ({ records }) => {
     }`,
   }));
 
-  const Item = ({ title, id }) => (
-    <View style={styles.item}>
-      <Text style={styles.title}>{title}</Text>
-      {showAll ? <Button title="Delete" 
-          onPress={()=>deleteRecord(id)}
-      /> : <Fragment />}
-    </View>
-  );
-  const renderItem = ({ item }) => {
-   return (<Item title={item.title} id={item.id} />);
-  };
-
-  const deleteRecord = async (id)=>{
+  const deleteRecord = async id => {
+    if (!window.confirm('You sure?')) {
+      return;
+    }
     console.log('delete ', id);
     const url = `${Constants.REST_ENDPOINT}record/${id}`;
     try {
       const response = await fetch(url, {
-        method: "DELETE", 
-        mode: "cors", 
-        cache: "no-cache", 
-        credentials: "same-origin",
-        redirect: "follow", 
-        referrerPolicy: "no-referrer", 
+        method: 'DELETE',
+        mode: 'cors',
+        cache: 'no-cache',
+        credentials: 'same-origin',
+        redirect: 'follow',
+        referrerPolicy: 'no-referrer',
       });
 
       if (response.ok) {
-        alert('Deleted ', id)
+        alert('Deleted ', id);
+        await onUpdate();
       } else {
-        console.log("Network response was not ok.");
+        console.log('Network response was not ok.');
       }
     } catch (error) {
-      alert("Error: " + error);
+      alert('Error: ' + error);
     }
-  }
+  };
+
+  const Item = ({ title, id }) => (
+    <View style={styles.item}>
+      <Text style={styles.title}>{title}</Text>
+      {showAll ? (
+        <Button title="Delete" onPress={() => deleteRecord(id)} />
+      ) : (
+        <Fragment />
+      )}
+    </View>
+  );
 
   return (
     <Fragment>
@@ -76,7 +79,7 @@ const RecordList = ({ records }) => {
       <View>
         <FlatList
           data={DATA}
-          renderItem={renderItem}
+          renderItem={({ item }) => <Item title={item.title} id={item.id} />}
           keyExtractor={item => item.id}
         />
       </View>
